Split up seed data file into seperate pieces
