+++ src/main/java/com/lambdaschool/tiemendo/model/CropType.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Entity
-@Table
+@Table(name="croptype")
 public class CropType {
 
     @Id
