+++ src/main/java/com/lambdaschool/tiemendo/repository/FarmerContactRepository.java
@@ -1,0 +1,7 @@
+package com.lambdaschool.tiemendo.repository;
+
+import com.lambdaschool.tiemendo.model.FarmerContact;
+import org.springframework.data.repository.CrudRepository;
+
+public interface FarmerContactRepository extends CrudRepository<FarmerContact, Long> {
+}
