+++ src/main/java/com/lambdaschool/tiemendo/repository/FarmerLocationRepository.java
@@ -1,0 +1,7 @@
+package com.lambdaschool.tiemendo.repository;
+
+import com.lambdaschool.tiemendo.model.FarmerLocation;
+import org.springframework.data.repository.CrudRepository;
+
+public interface FarmerLocationRepository extends CrudRepository<FarmerLocation, Long> {
+}
